# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: aboutme API
  version: 0.1.0
  description: |
    `/api/v1` — the REST contract for the aboutme resume builder + hosted
    display service. This document is the single source of truth for the
    API. Contract changes update this file instead of redefining shapes
    elsewhere. Go, TypeScript, and Dart clients are generated from it.

    ## Response envelope

    Every response body is one of:

    - Success: `Envelope` — `{"data": ...}`.
    - Failure: `Error` — `{"error": {"code": "...", "message": "..."}}`.
      Both `code` and `message` are always present.

    ## Write safety (revisions, `If-Match`, idempotency)

    Every resume is guarded by a monotonically increasing `revision`
    (server-assigned bigint). Because 64-bit integers are not safely
    representable in JSON numbers for TS/Dart clients, `revision` is always
    serialized as a **string** — see the `Revision` schema.

    - **Optimistic concurrency.** Every mutation of an existing resume must
      carry the `If-Match` header in ETag form:
      `If-Match: "r<revision>"` (see the `IfMatch` parameter). Creating a
      resume has no prior revision and rejects `If-Match`. Authentication and
      session mutations do not use the resume precondition.
    - **Staleness is `412`, never `409`.** When the supplied `If-Match` does
      not match the resource's current revision, the server returns
      `412 Precondition Failed` with an `Error` body and the current
...
  license:
...
servers:
...
paths:
...
  /auth/github/start:
    get:
      operationId: getAuthGitHubStart
      tags: [auth]
      security: []
      summary: Begin "Sign in with GitHub"
      description: >-
        Begins a fresh OAuth2 login transaction (PKCE S256; GitHub has no
        OIDC ID token, so no nonce) and redirects the browser to GitHub's
        own authorize endpoint. Sets the `__Host-oauth-tx` cookie (opaque
        transaction handle, 10-minute TTL) that the matching `GET
        /auth/github/callback` consumes. `HEAD` is rejected with `405`.
        `GET` accepts only login; `link` and `reauth` return `405` with
        `Allow: POST` before any transaction is created.
      parameters:
        - $ref: "#/components/parameters/AuthPurpose"
      responses:
        "302":
          description: Success — a redirect to GitHub's own authorize endpoint.
          headers:
            Location:
              schema:
                type: string
                format: uri
              example: https://github.com/login/oauth/authorize?...
        "405":
          description: >-
            Either an unsupported method, including `HEAD`, or a `GET`
            carrying `?purpose=link` or `?purpose=reauth`. The latter
            purposes are served only by `POST` and return `Allow: POST`.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429":
          $ref: "#/components/responses/AuthStartRateLimited"
    post:
      operationId: postAuthGitHubStart
      tags: [auth]
      security:
        - sessionCookie: []
          csrfToken: []
      summary: Begin a GitHub link or reauthentication
      description: >-
        Begins an OAuth transaction for the CALLER's own already-
        authenticated account — `purpose=link` (attach this provider
        identity to that account) or `purpose=reauth` (refresh the
        session's `reauthenticated_at` against an ALREADY-linked
        identity) — and returns GitHub's authorize URL for the client to
        navigate to. Sets the same `__Host-oauth-tx` cookie (opaque
        transaction handle, 10-minute TTL) the matching `GET
        /auth/github/callback` consumes.


        The response deliberately does NOT redirect: a `302` answering a
        `fetch()` would be followed by the fetch, and the provider's
        consent screen has to be a real top-level navigation the client
        performs itself with the returned URL.


        Link and reauthentication starts are privileged account actions.
        They therefore use `POST` behind the normal session, CSRF token,
        and exact-Origin checks. `GET` refuses both purposes.
      parameters:
        - $ref: "#/components/parameters/AuthLinkPurpose"
      responses:
        "200":
          description: The authorize URL to navigate the browser to.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthStartResponse"
              example:
                data:
                  authorizeUrl: https://github.com/login/oauth/authorize?...
        "400":
          description: >-
            `purpose` was absent or was something other than `link` or
            `reauth`. This route serves exactly those two — a login start
            is the `GET`'s job — and never substitutes a default, which
            would hand back an authorize URL for a flow the caller did not
            ask for.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error:
                  code: bad_request
                  message: purpose must be link or reauth
        "401":
          description: No valid `__Host-session` cookie. Clears the cookie.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error:
                  code: session_required
                  message: a valid session is required
        "403":
          description: >-
            Either the CSRF chain rejected the request (missing/mismatched
            `Origin`, missing/wrong `csrfToken`) or — `purpose=link` only —
            the session's last full OAuth login is more than 15 minutes
            old. The reauth check runs BEFORE any transaction row is
            created, so a stale caller never gets a live transaction
            handle. `purpose=reauth` has no such gate: its whole point is
            to establish recency.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                csrf_rejected:
                  summary: Failed the CSRF chain
                  value:
                    error:
                      code: csrf_rejected
                      message: CSRF validation failed
                reauth_required:
                  summary: "purpose=link, stale reauth window"
                  value:
                    error:
                      code: reauth_required
                      message: recent reauthentication is required
        "429":
          $ref: "#/components/responses/AuthStartRateLimited"
  /auth/github/callback:
    get:
      operationId: getAuthGitHubCallback
      tags: [auth]
      security: []
      summary: Complete "Sign in with GitHub"
      description: >-
        Consumes the pending OAuth transaction (`__Host-oauth-tx`
        cookie), exchanges the authorization code, and — for
        `purpose=login` only — fetches the authenticated user's verified
        primary email from GitHub's REST API (AC-AUTH-003) — deliberately
        plain OAuth2, never OIDC: no ID token, no issuer/audience/
        signature/nonce to verify. `purpose=link`/`reauth` never fetches
        email at all (no email check for linking — see below). Otherwise
        follows the exact same purpose-dependent resolution and
        always-redirect contract as `GET /auth/google/callback` — see that
        operation's own description
        for the full `purpose=login` vs `purpose=link`/`reauth` behavior
        and redirect targets.
      parameters:
        - name: code
          in: query
          required: false
          schema:
            type: string
          description: The provider's authorization code (RFC 6749 §4.1.2).
            Absent on a provider-signaled rejection (see `error` below).
        - name: state
          in: query
          required: false
          schema:
            type: string
          description: Must match the pending transaction's own `state` (RFC
            6749 §10.12) or the callback is rejected with `auth_failed`.
        - name: error
          in: query
          required: false
          schema:
            type: string
          description: Set by the PROVIDER itself when the visitor declines
            consent — `error=access_denied`, mapped to this server's own
            `cancelled` code.
      responses:
        "302":
          description: Always redirects — see the operation description for
            the success/rejected branches.
          headers:
            Location:
              schema:
                type: string
                format: uri
              example: https://aboutme.vn/login?error=auth_failed
...
components:
  responses:
    AuthStartRateLimited:
      description: >-
        The start routes' own rate limit is exhausted. Anonymous sign-in
        `GET` starts are keyed by client IP. Authenticated link and
        reauthentication `POST` starts are keyed by `(account, client IP)`.
        Both are tighter than the whole-server default because each accepted
        start writes a database row. Carries `Retry-After` in whole seconds.
      headers:
        Retry-After:
          schema:
            type: integer
            minimum: 1
          description: Whole seconds to wait before retrying.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error:
              code: rate_limited
              message: too many requests; retry later
  schemas:
...
    AuthStartResponse:
      type: object
      description: >-
        `POST /auth/{provider}/start`'s success envelope: the provider
        authorize URL the client navigates the browser to itself.
      required:
        - data
      properties:
        data:
          type: object
          required:
            - authorizeUrl
          properties:
            authorizeUrl:
              type: string
              format: uri
              description: >-
                The provider's own authorize endpoint, carrying this
                transaction's `state` and PKCE challenge. Single use: it
                is bound to the `__Host-oauth-tx` cookie set on the same
                response, and both expire together after 10 minutes.
              example: https://accounts.google.com/o/oauth2/v2/auth?...
    Error:
      type: object
      description: Failure response wrapper.
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
              description: Stable machine-readable error code (e.g.
                `slug_taken`, `precondition_failed`).
              example: precondition_failed
            message:
              type: string
              description: Human-readable message; not intended to be
                parsed by clients.
              example: The resume was modified since you last loaded it.
...
  parameters:
    AuthPurpose:
      name: purpose
      in: query
      required: false
      description: >-
        Why this OAuth transaction is being started. `GET` serves only
        `login`, which is also the default. An absent or unrecognized value
        is treated as `login`; `link` and `reauth` are the only exceptions
        and return `405` with `Allow: POST`.


        - `login` (default): an ordinary, unauthenticated sign-in/
          registration attempt. Reachable from anywhere — a bookmarked or
          shared link, another site's "continue with aboutme" button —
          with no session, CSRF, or same-site requirement at all.

        - `link` / `reauth`: **not served by `GET`**. Both
          are `405` here, with `Allow: POST`, before any session read,
          database access, or cookie is set — see the `POST` operation on
          the same path, and `AuthLinkPurpose` below.


        The unconditional refusal prevents a cross-site page from chaining
        reauthentication and linking through browser navigations. Privileged
        starts use the CSRF-protected `POST` operation instead.
      schema:
        type: string
        enum: [login]
        default: login
      example: login
    AuthLinkPurpose:
      name: purpose
      in: query
      required: true
      description: >-
        Which privileged flow `POST /auth/{provider}/start` is beginning
        for the caller's own already-authenticated account. Required, with
        no default: this route serves exactly these two, and substituting
        one would return an authorize URL for a flow the caller did not
        ask for.


        - `link`: attach this provider identity to the caller's account.
          Additionally requires a session whose last full OAuth login
          (`reauthenticated_at`) is within the last 15 minutes, checked
          BEFORE any transaction row is created — `403 reauth_required`
          otherwise.

        - `reauth`: refresh the caller's own session's
          `reauthenticated_at` via a fresh round trip against an
          ALREADY-linked provider (the compensating control the long
          idle and absolute session timeouts require before a sensitive
          operation). Deliberately does NOT itself require a recent
          reauthentication — that would be circular, since its whole point
          is to establish one. What makes that safe is that this route is
          `csrfToken`-protected, so a cross-site page cannot make the
          request at all.
      schema:
        type: string
        enum: [link, reauth]
      example: link
...
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: __Host-session
      description: >-
        Server-side opaque session token: `httpOnly`,
        `Secure`, `SameSite=Lax`, no `Domain` attribute (`__Host-`
        prefix). Never a JWT — a revocable, Postgres-backed session row.
        Every non-OAuth-redirect operation in the `auth` tag requires
        this.
    csrfToken:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: >-
        Synchronizer CSRF token:
        `base64url` (no padding) encoding of the authenticated session's
        own CSRF secret, obtained from `GET /me`'s `csrfToken` field —
        the only place it is ever returned — and required, alongside
        `sessionCookie`, on every mutating cookie-authenticated request
        (together with `SameSite=Lax` and an exact `Origin`/`Referer`
        check). Declared together with `sessionCookie` (AND, not OR) on
        every mutating operation in the `auth` tag.